% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_tables.R
\name{rankings_table}
\alias{rankings_table}
\title{Create rankings table}
\usage{
rankings_table(input_df, input_position, input_player, input_rankers)
}
\arguments{
\item{input_df}{dataframe of rankings}

\item{input_position}{position to filter}

\item{input_player}{players to filter}

\item{input_rankers}{rankers to filter}
}
\value{
dataframe of rankings
}
\description{
Create rankings table
}
